+++ packages/filesystem/src/browser/breadcrumbs/filepath-breadcrumbs-contribution.ts
@@ -14,18 +14,21 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-import { BreadcrumbsContribution } from '@theia/core/lib/browser/breadcrumbs/breadcrumbs-contribution';
-import { Breadcrumb } from '@theia/core/lib/browser/breadcrumbs/breadcrumb';
-import { FilepathBreadcrumb } from './filepath-breadcrumb';
-import { injectable, inject } from 'inversify';
-import { LabelProvider, Widget } from '@theia/core/lib/browser';
-import { FileSystem, FileStat } from '../../common';
+import { Disposable, Emitter, Event } from '@theia/core';
+import { injectable, inject } from '@theia/core/shared/inversify';
 import URI from '@theia/core/lib/common/uri';
+import { Breadcrumb, BreadcrumbsContribution, CompositeTreeNode, LabelProvider, SelectableTreeNode, Widget } from '@theia/core/lib/browser';
+import { FilepathBreadcrumb } from './filepath-breadcrumb';
 import { BreadcrumbsFileTreeWidget } from './filepath-breadcrumbs-container';
 import { DirNode } from '../file-tree';
-import { Disposable } from '@theia/core';
+import { FileService } from '../file-service';
+import { FileStat } from '../../common/files';
 
 export const FilepathBreadcrumbType = Symbol('FilepathBreadcrumb');
+
+export interface FilepathBreadcrumbClassNameFactory {
+    (location: URI, index: number): string;
+}
 
 @injectable()
 export class FilepathBreadcrumbsContribution implements BreadcrumbsContribution {
@@ -33,11 +36,16 @@ export class FilepathBreadcrumbsContribution implements BreadcrumbsContribution 
     @inject(LabelProvider)
     protected readonly labelProvider: LabelProvider;
 
-    @inject(FileSystem)
-    protected readonly fileSystem: FileSystem;
+    @inject(FileService)
+    protected readonly fileSystem: FileService;
 
     @inject(BreadcrumbsFileTreeWidget)
     protected readonly breadcrumbsFileTreeWidget: BreadcrumbsFileTreeWidget;
+
+    protected readonly onDidChangeBreadcrumbsEmitter = new Emitter<URI>();
+    get onDidChangeBreadcrumbs(): Event<URI> {
+        return this.onDidChangeBreadcrumbsEmitter.event;
+    }
 
     readonly type = FilepathBreadcrumbType;
     readonly priority: number = 100;
@@ -46,13 +54,30 @@ export class FilepathBreadcrumbsContribution implements BreadcrumbsContribution 
         if (uri.scheme !== 'file') {
             return [];
         }
-        return (await Promise.all(uri.allLocations.reverse()
-            .map(async u => new FilepathBreadcrumb(
-                u,
-                this.labelProvider.getName(u),
-                this.labelProvider.getLongName(u),
-                await this.labelProvider.getIcon(u) + ' file-icon'
-            )))).filter(b => this.filterBreadcrumbs(uri, b));
+        const getContainerClass = this.getContainerClassCreator(uri);
+        const getIconClass = this.getIconClassCreator(uri);
+        return uri.allLocations
+            .map((location, index) => {
+                const icon = getIconClass(location, index);
+                const containerClass = getContainerClass(location, index);
+                return new FilepathBreadcrumb(
+                    location,
+                    this.labelProvider.getName(location),
+                    this.labelProvider.getLongName(location),
+                    icon,
+                    containerClass,
+                );
+            })
+            .filter(b => this.filterBreadcrumbs(uri, b))
+            .reverse();
+    }
+
+    protected getContainerClassCreator(fileURI: URI): FilepathBreadcrumbClassNameFactory {
+        return (location, index) => location.isEqual(fileURI) ? 'file' : 'folder';
+    }
+
+    protected getIconClassCreator(fileURI: URI): FilepathBreadcrumbClassNameFactory {
+        return (location, index) => location.isEqual(fileURI) ? this.labelProvider.getIcon(location) + ' file-icon' : '';
     }
 
     protected filterBreadcrumbs(_: URI, breadcrumb: FilepathBreadcrumb): boolean {
@@ -63,30 +88,42 @@ export class FilepathBreadcrumbsContribution implements BreadcrumbsContribution 
         if (!FilepathBreadcrumb.is(breadcrumb)) {
             return undefined;
         }
-        const folderFileStat = await this.fileSystem.getFileStat(breadcrumb.uri.parent.toString());
+        const folderFileStat = await this.fileSystem.resolve(breadcrumb.uri.parent);
         if (folderFileStat) {
             const rootNode = await this.createRootNode(folderFileStat);
-            await this.breadcrumbsFileTreeWidget.model.navigateTo(rootNode);
-            Widget.attach(this.breadcrumbsFileTreeWidget, parent);
-            return {
-                dispose: () => {
-                    // Clear model otherwise the next time a popup is opened the old model is rendered first
-                    // and is shown for a short time period.
-                    this.breadcrumbsFileTreeWidget.model.root = undefined;
-                    Widget.detach(this.breadcrumbsFileTreeWidget);
-                }
-            };
+            if (rootNode) {
+                const { model } = this.breadcrumbsFileTreeWidget;
+                await model.navigateTo({ ...rootNode, visible: false });
+                Widget.attach(this.breadcrumbsFileTreeWidget, parent);
+                const toDisposeOnTreePopulated = model.onChanged(() => {
+                    if (CompositeTreeNode.is(model.root) && model.root.children.length > 0) {
+                        toDisposeOnTreePopulated.dispose();
+                        const targetNode = model.getNode(breadcrumb.uri.path.toString());
+                        if (targetNode && SelectableTreeNode.is(targetNode)) {
+                            model.selectNode(targetNode);
+                        }
+                        this.breadcrumbsFileTreeWidget.activate();
+                    }
+                });
+                return {
+                    dispose: () => {
+                        // Clear model otherwise the next time a popup is opened the old model is rendered first
+                        // and is shown for a short time period.
+                        toDisposeOnTreePopulated.dispose();
+                        this.breadcrumbsFileTreeWidget.model.root = undefined;
+                        Widget.detach(this.breadcrumbsFileTreeWidget);
+                    }
+                };
+            }
         }
     }
 
     protected async createRootNode(folderToOpen: FileStat): Promise<DirNode | undefined> {
-        const folderUri = new URI(folderToOpen.uri);
+        const folderUri = folderToOpen.resource;
         const rootUri = folderToOpen.isDirectory ? folderUri : folderUri.parent;
-        const name = this.labelProvider.getName(rootUri);
-        const rootStat = await this.fileSystem.getFileStat(rootUri.toString());
+        const rootStat = await this.fileSystem.resolve(rootUri);
         if (rootStat) {
-            const label = await this.labelProvider.getIcon(rootStat);
-            return DirNode.createRoot(rootStat, name, label);
+            return DirNode.createRoot(rootStat);
         }
     }
 }