+++ packages/workspace/src/browser/workspace-breadcrumbs-contribution.ts
@@ -15,8 +15,8 @@
  ********************************************************************************/
 
 import { FilepathBreadcrumb } from '@theia/filesystem/lib/browser/breadcrumbs/filepath-breadcrumb';
-import { FilepathBreadcrumbsContribution } from '@theia/filesystem/lib/browser/breadcrumbs/filepath-breadcrumbs-contribution';
-import { inject, injectable } from 'inversify';
+import { FilepathBreadcrumbClassNameFactory, FilepathBreadcrumbsContribution } from '@theia/filesystem/lib/browser/breadcrumbs/filepath-breadcrumbs-contribution';
+import { inject, injectable } from '@theia/core/shared/inversify';
 import { WorkspaceService } from './workspace-service';
 import URI from '@theia/core/lib/common/uri';
 
@@ -26,8 +26,31 @@ export class WorkspaceBreadcrumbsContribution extends FilepathBreadcrumbsContrib
     @inject(WorkspaceService)
     protected readonly workspaceService: WorkspaceService;
 
+    getContainerClassCreator(fileURI: URI): FilepathBreadcrumbClassNameFactory {
+        const workspaceRoot = this.workspaceService.getWorkspaceRootUri(fileURI);
+        return (location, index) => {
+            if (location.isEqual(fileURI)) {
+                return 'file';
+            } else if (workspaceRoot?.isEqual(location)) {
+                return 'root_folder';
+            }
+            return 'folder';
+        };
+    }
+
+    getIconClassCreator(fileURI: URI): FilepathBreadcrumbClassNameFactory {
+        const workspaceRoot = this.workspaceService.getWorkspaceRootUri(fileURI);
+        return (location, index) => {
+            if (location.isEqual(fileURI) || workspaceRoot?.isEqual(location)) {
+                return this.labelProvider.getIcon(location) + ' file-icon';
+            }
+            return '';
+        };
+    }
+
     protected filterBreadcrumbs(uri: URI, breadcrumb: FilepathBreadcrumb): boolean {
         const workspaceRootUri = this.workspaceService.getWorkspaceRootUri(uri);
-        return super.filterBreadcrumbs(uri, breadcrumb) && (!workspaceRootUri || !breadcrumb.uri.isEqualOrParent(workspaceRootUri));
+        const firstCrumbToHide = this.workspaceService.isMultiRootWorkspaceOpened ? workspaceRootUri?.parent : workspaceRootUri;
+        return super.filterBreadcrumbs(uri, breadcrumb) && (!firstCrumbToHide || !breadcrumb.uri.isEqualOrParent(firstCrumbToHide));
     }
 }